Portability fix for printing pid
We cannot rely on "%d" to print the pid correctly; POSIX specifies that:
    pid_t... shall be signed integer types.
    https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_types.h.html

That page also says:
    The implementation shall support one or more programming
    environments in which the widths of ... pid_t, ...  are no greater
    than the width of type long.
but if we were to rely on that, we'd need to check that we were in the
relevant programming environment.  Easier just to use "%jd".

+++ libcperciva/util/daemonize.c
@@ -1,4 +1,5 @@
 #include <errno.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <unistd.h>
 
@@ -53,7 +54,7 @@ writepid(const char * spid)
 		warnp("fopen(%s)", spid);
 		goto err0;
 	}
-	if (fprintf(f, "%d", getpid()) < 0) {
+	if (fprintf(f, "%jd", (intmax_t)getpid()) < 0) {
 		warnp("fprintf");
 		goto err1;
 	}
